[WIP] refactor backend, splitted apps and added tournaments app
