+++ backend/app/settings.py
@@ -71,6 +71,7 @@ INSTALLED_APPS = [
     'rest_framework',
     'rest_framework_simplejwt',
     'app.users',
+    'app.games',
     'app',
     'channels',
     'corsheaders',
