+++ backend/app/users/admin.py
@@ -1,0 +1,5 @@
+from django.contrib import admin
+from app.users.models import CustomUser
+
+# Register your models here.
+admin.site.register(CustomUser)
